avance de reportes

+++ Unna.OperationalReport.WebSite/Controllers/Admin/IngenieroProceso/Reporte/Diario/BoletaVentaGnsController.cs
@@ -7,6 +7,7 @@ using Unna.OperationalReport.Service.Reportes.ReporteDiario.BoletaCnpc.Dtos;
 using Unna.OperationalReport.Service.Reportes.ReporteDiario.BoletaVentaGns.Dtos;
 using Unna.OperationalReport.Service.Reportes.ReporteDiario.BoletaVentaGns.Servicios.Abstracciones;
 using Unna.OperationalReport.Tools.Comunes.Infraestructura.Dtos;
+using Unna.OperationalReport.Tools.Comunes.Infraestructura.Utilitarios;
 using Unna.OperationalReport.Tools.Seguridad.Servicios.General.Dtos;
 using Unna.OperationalReport.Tools.WebComunes.ApiWeb.Auth.Atributos;
 using Unna.OperationalReport.Tools.WebComunes.WebSite.Base;
@@ -53,37 +54,73 @@ namespace Unna.OperationalReport.WebSite.Controllers.Admin.IngenieroProceso.Repo
         }
 
         [HttpGet("GenerarExcel")]
+        [RequiereAcceso()]
         public async Task<IActionResult> GenerarExcelAsync()
         {
-            var operativo = await _boletaVentaGnsServicio.ObtenerAsync(ObtenerIdUsuarioActual() ?? 0);
-            if (!operativo.Completado || operativo.Resultado == null)
+            string? url = await GenerarAsync();
+            if (string.IsNullOrWhiteSpace(url))
             {
                 return File(new byte[0], "application/octet-stream");
             }
-            var dato = operativo.Resultado;
-                 
-            var tempFilePath = $"{_general.RutaArchivos}{Guid.NewGuid()}.xlsx";
+            var bytes = System.IO.File.ReadAllBytes(url);
+            System.IO.File.Delete(url);
 
+            string nombreArchivo = FechasUtilitario.ObtenerFechaSegunZonaHoraria(DateTime.UtcNow).ToString("dd-MM-yyyy HH:mm:ss tt");
+            return File(bytes, "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", $"BOLETA DE VENTA DEL GAS NATURAL SECO DE UNNA LOTE IV A ENEL - {nombreArchivo}.xlsx");
+        }
+
+        [HttpGet("GenerarPdf")]
+        [RequiereAcceso()]
+        public async Task<IActionResult> GenerarPdfAsync()
+        {
+            string? url = await GenerarAsync();
+            if (string.IsNullOrWhiteSpace(url))
+            {
+                return File(new byte[0], "application/octet-stream");
+            }
+            var tempFilePathPdf = $"{_general.RutaArchivos}{Guid.NewGuid()}.pdf";
+            var workbook = new Workbook(url);
+            workbook.Save(tempFilePathPdf);
+            var bytes = System.IO.File.ReadAllBytes(tempFilePathPdf);
+            System.IO.File.Delete(url);
+            System.IO.File.Delete(tempFilePathPdf);
+            string nombreArchivo = FechasUtilitario.ObtenerFechaSegunZonaHoraria(DateTime.UtcNow).ToString("dd-MM-yyyy HH:mm:ss tt");
+
+            return File(bytes, "application/pdf", $"BOLETA DE VENTA DEL GAS NATURAL SECO DE UNNA LOTE IV A ENEL - {nombreArchivo}.pdf");
+        }
+
+
+        private async Task<string?> GenerarAsync()
+        {
+            var operativo = await _boletaVentaGnsServicio.ObtenerAsync(ObtenerIdUsuarioActual() ?? 0);
+            if (!operativo.Completado || operativo.Resultado == null || operativo.Resultado.General == null)
+            {
+                return null;
+            }
+            var dato = operativo.Resultado;
+
+            object cuerpo = new
+            {
+                NombreReporte = dato.General.NombreReporte,
+                Compania = dato.General.Nombre,
+                Version = $"Versión {dato.General.Version}",
+                Fecha = dato.Fecha,
+                Mpcs = dato.Mpcs,
+                BtuPcs = dato.BtuPcs,
+                Mmbtu = dato.Mmbtu,
+                RazonSocial = dato.Empresa,
+                Abreviatura = dato.Abreviatura,
+            };
+            var tempFilePath = $"{_general.RutaArchivos}{Guid.NewGuid()}.xlsx";
             using (var template = new XLTemplate($"{_hostingEnvironment.WebRootPath}\\plantillas\\reporte\\diario\\BoletaDeVentaGns.xlsx"))
             {
-                template.AddVariable(dato);
+                template.AddVariable(cuerpo);
                 template.Generate();
                 template.SaveAs(tempFilePath);
             }
-           
-            var tempFilePathPdf = $"{_general.RutaArchivos}{Guid.NewGuid()}.pdf";
-            var workbook = new Workbook(tempFilePath);
-            workbook.Save(tempFilePathPdf);
-
-            var bytes = System.IO.File.ReadAllBytes(tempFilePathPdf);
-
-
-            System.IO.File.Delete(tempFilePath);
-            System.IO.File.Delete(tempFilePathPdf);
-
-            return File(bytes, "application/pdf", $"BoletaVentaGns-{dato.Fecha.Replace("/", "-")}.pdf");
-            //return File(bytes, "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", $"BoletaCnpc-{dato.Fecha.Replace("/", "-")}.xlsx");
+            return tempFilePath;
         }
+
 
         [HttpGet("Pdf")]
         public async Task<IActionResult> Pdf()
